ACPENG-2476 Allow Oracle charset to be specified
This adds a new argument to the module to permit two Oracle (and MSSQL)
fields to be overridden from the defaults.

I overengineered this earlier in development by writing something like:

  startswith(var.engine_type, "oracle") ? var.db_character_set : null

However, that proved unnecessary as you can just default the variable to
be `null` unless specified. If it's Oracle and the variable is unset,
it'll simply pick the default from upstream; if it's MySQL/Postgres and
the variable is unset, upstream just ignores it as it's null.