Use try() to prevent null error
It turns out coalesce() will crash out if it ends up being unable to
return a value that isn't `null`, but we want the computer to return
`null` if the variable is unset and the storage_type isn't `gp3`.

However, we want it to crash if the storage_type *is* `gp3` but the
IOPS is, after all our checks, still undefined.

This commit wraps this in a `try()` function which evaluates all its
argument expressions in turn and returns the result of the first one
that does not produce any errors. As such, if `gp2` is set, the inner
`coalesce()` will produce an error, and the computer will proceed onto
the next argument and eventually return `null`.