Add lifecycle wrapper around the `iops` value
I think there is a bug in the version of the (upstream) AWS provider
we're using which is converting `null` to `0`, meaning that although if
we create a small RDS with `iops` set to `null` AWS will convert that
under-the-hood to `3000`, if we then subsequently modify that RDS
Terraform tries to re-set that `3000` to `0`, which is not allowed.

This commit wraps the `iops` argument in the `lifecycle` meta-argument,
specifying that Terraform should `ignore_changes` to this value. This
should stop Terraform noticing the IOPS has (from its perspective)
"changed", thereby preventing Terraform from "fixing" the divergence.

I wanted to wrap this in an if statement that toggles `ignore_changes`
on if we're using the fallback IOPS computation in `locals`, but
Terraform doesn't let you do dynamic things inside a `lifecycle` block.

See also 2dda3af822594c93dec7bc10dbcad5a2c33a31df